event handler for actions are being controlled
